Fix implicit casts of numbers & int multiplication
* Cast or replace with appropriate type to prevent overflow or data loss

+++ core/src/main/java/org/apache/accumulo/core/clientImpl/ReplicationClient.java
@@ -56,7 +56,7 @@ public class ReplicationClient {
       log.debug("Could not get ReplicationCoordinator connection to {}, will retry",
           context.getInstanceName());
       try {
-        Thread.sleep(attempts * 250);
+        Thread.sleep((long) attempts * 250);
       } catch (InterruptedException e) {
         throw new AccumuloException(e);
       }
